Dry up custom immutable Map.Entry implementations
Follow-up to #88815.
No need to have two equivalent implementations here.
