+++ server/src/main/java/org/elasticsearch/cluster/DiffableUtils.java
@@ -13,6 +13,7 @@ import org.elasticsearch.common.collect.ImmutableOpenMap;
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
 import org.elasticsearch.common.io.stream.Writeable.Reader;
+import org.elasticsearch.common.util.Maps;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -153,7 +154,7 @@ public final class DiffableUtils {
                 inserts++;
             } else if (entry.getValue().equals(previousValue) == false) {
                 if (valueSerializer.supportsDiffableValues()) {
-                    diffs.add(mapEntry(entry.getKey(), valueSerializer.diff(entry.getValue(), previousValue)));
+                    diffs.add(new Maps.ImmutableEntry<>(entry.getKey(), valueSerializer.diff(entry.getValue(), previousValue)));
                 } else {
                     upserts.add(entry);
                 }
@@ -307,14 +308,14 @@ public final class DiffableUtils {
             for (int i = 0; i < diffsCount; i++) {
                 K key = keySerializer.readKey(in);
                 Diff<T> diff = valueSerializer.readDiff(in, key);
-                diffs.add(mapEntry(key, diff));
+                diffs.add(new Maps.ImmutableEntry<>(key, diff));
             }
             int upsertsCount = in.readVInt();
             upserts = upsertsCount == 0 ? List.of() : new ArrayList<>(upsertsCount);
             for (int i = 0; i < upsertsCount; i++) {
                 K key = keySerializer.readKey(in);
                 T newValue = valueSerializer.read(in, key);
-                upserts.add(mapEntry(key, newValue));
+                upserts.add(new Maps.ImmutableEntry<>(key, newValue));
             }
             this.builderCtor = builderCtor;
         }
@@ -400,25 +401,6 @@ public final class DiffableUtils {
                 }
             }
         }
-    }
-
-    private static <K, T> Map.Entry<K, T> mapEntry(K key, T newValue) {
-        return new Map.Entry<>() {
-            @Override
-            public K getKey() {
-                return key;
-            }
-
-            @Override
-            public T getValue() {
-                return newValue;
-            }
-
-            @Override
-            public T setValue(T value) {
-                throw new UnsupportedOperationException();
-            }
-        };
     }
 
     /**
